Ejercicio 1 - Completo

+++ Tema4/src/parte3/ejercicio1/Suma.java
@@ -3,19 +3,19 @@ package parte3.ejercicio1;
 public class Suma {
 
 	/**
-	 * Esta función recibe dos números y se encarga de calcular la suma de dichos
-	 * números.
+	 * Esta función recibe dos números enteros y se encarga de calcular la suma de
+	 * dichos números.
 	 * 
 	 * @param num1 El primer número entero pasado como parametro.
 	 * @param num2 El segundo número entero pasado como parametro.
 	 * @return La suma de los dos números.
 	 */
 	public int suma(int num1, int num2) {
-		// Creamos la variable sumaNumeros como int para almmacenar el valor de las dos
-		// sumas pasadas como parametro.
+		// Creamos la variable sumaNumeros como int para almacenar el valor de la suma
+		// de los dos números pasados como parametro.
 		int sumaNumeros;
 
-		// Realizamos la ecuación y lo alamcenamos en sumaNumeros.
+		// Realizamos la ecuación y lo almacenamos en sumaNumeros.
 		sumaNumeros = num1 + num2;
 
 		// Devolvemos la variable sumaNumeros donde se almacena el resultado de la suma
@@ -23,22 +23,23 @@ public class Suma {
 	}
 
 	/**
-	 * Esta función recibe dos números y se encarga de calcular la suma de dichos
-	 * números.
+	 * Esta función recibe dos números decimales y se encarga de calcular la suma de
+	 * dichos números.
 	 * 
-	 * @param num1 El primer número pasado como parametro.
-	 * @param num2 El segundo número pasado como parametro.
-	 * @return La suma de los dos números.F
+	 * @param num1 El primer número decimal pasado como parametro.
+	 * @param num2 El segundo número decimal pasado como parametro.
+	 * @return La suma de los dos números.
 	 */
-	public int suma(int num1, int num2) {
-		// Creamos la variable sumaNumeros como int para almmacenar el valor de las dos
-		// sumas pasadas como parametro.
-		int sumaNumeros;
+	public double suma(double num1, double num2) {
+		// Creamos la variable sumaNumeros como double para almacenar el valor de la
+		// suma de los dos números pasados como parametro.
+		double sumaNumeros;
 
-		// Realizamos la ecuación y lo alamcenamos en sumaNumeros.
+		// Realizamos la ecuación y lo almacenamos en sumaNumeros.
 		sumaNumeros = num1 + num2;
 
-		// Devolvemos la variable sumaNumeros donde se almacena el resultado de la suma
+		// Devolvemos la variable sumaNumeros donde se almacena el resultado de la suma.
 		return sumaNumeros;
 	}
+
 }
